Normalise postcodes with JavaScript
This adds JavaScript which intercepts the submission of the Coronavirus
local restriction postcode form and converts an input that looks like a
postcode into a formatted approach (uppercase with a space between
inward and onward codes).

This has been done to increase the chance that the same postcodes
entered produce cachable HTTP responses. It is intended as a less aggressive
replacement to our CDN code [1] which will produce a nicer user
experience.

To avoid conflicting with the postcode validation in the app I've taken
a reasonably lax approach to checking for a postcode. This just checks
and normalises something that looks like a common UK postcode. It leaves
erroneous input intact which should be beneficial in helping users
correct their errors.

[1]: https://github.com/alphagov/govuk-cdn-config/pull/312/files#diff-94c0786370f6c095cf75ced830bd83a978ee352c1a5724c223f812f256699350R254-R265