Fix failing JS feature test
The lack of a title attribute in this stub was causing an exception to
occur when a JSON view of a browse page was being rendered (no method
`sub` for nil).

Until the previous commit this test must have not actually been
succeeding with JavaScript and merely succeeding due to no-JS fallbacks.
This indicates that this test isn't sufficient for the scenario it is
testing. I'm not too sure how to fix that now though so I'm merely doing
enough to resolve the problem.